Split up benchmark_fstream.cpp
