+++ test/test_fstream.cpp
@@ -10,12 +10,8 @@
 #include <boost/nowide/fstream.hpp>
 #include <boost/nowide/cstdio.hpp>
 #include <boost/nowide/convert.hpp>
-#define BOOST_CHRONO_HEADER_ONLY
-#include <boost/chrono.hpp>
 #include <fstream>
 #include <iostream>
-#include <iomanip>
-#include <vector>
 #include "test.hpp"
 
 #ifdef BOOST_MSVC
@@ -23,129 +19,6 @@
 #endif
 
 namespace nw = boost::nowide;
-template<typename FStream>
-class io_fstream
-{
-public:
-    void open(char const *file)
-    {
-        f_.open(file, std::fstream::out | std::fstream::in | std::fstream::trunc);
-        TEST(f_);
-    }
-    void write(char *buf, int size)
-    {
-        f_.write(buf, size);
-    }
-    void read(char *buf, int size)
-    {
-        f_.read(buf, size);
-    }
-    void rewind()
-    {
-        f_.seekg(0);
-        f_.seekp(0);
-    }
-    void flush()
-    {
-        f_ << std::flush;
-    }
-    void close()
-    {
-        f_.close();
-    }
-
-private:
-    FStream f_;
-};
-
-class io_stdio
-{
-public:
-    void open(char const *file)
-    {
-        f_ = fopen(file, "w+");
-        TEST(f_);
-    }
-    void write(char *buf, int size)
-    {
-        fwrite(buf, 1, size, f_);
-    }
-    void read(char *buf, int size)
-    {
-        size_t res = fread(buf, 1, size, f_);
-        (void)res;
-    }
-    void rewind()
-    {
-        ::rewind(f_);
-    }
-    void flush()
-    {
-        fflush(f_);
-    }
-    void close()
-    {
-        fclose(f_);
-        f_ = 0;
-    }
-
-private:
-    FILE *f_;
-};
-
-template<typename FStream>
-void test_io(char *file, char const *type)
-{
-    std::cout << "Testing I/O performance " << type << std::endl;
-    FStream tmp;
-    tmp.open(file);
-    int data_size = 64 * 1024 * 1024;
-    for(int block_size = 16; block_size <= 8192; block_size *= 2)
-    {
-        std::vector<char> buf(block_size, ' ');
-        int size = 0;
-        tmp.rewind();
-        boost::chrono::high_resolution_clock::time_point t1 = boost::chrono::high_resolution_clock::now();
-        while(size < data_size)
-        {
-            tmp.write(&buf[0], block_size);
-            size += block_size;
-        }
-        tmp.flush();
-        boost::chrono::high_resolution_clock::time_point t2 = boost::chrono::high_resolution_clock::now();
-        double tm = boost::chrono::duration_cast<boost::chrono::milliseconds>(t2 - t1).count() * 1e-3;
-        // heatup
-        if(block_size >= 32)
-            std::cout << "  write block size " << std::setw(8) << block_size << " " << std::fixed << std::setprecision(3)
-                      << (data_size / 1024.0 / 1024 / tm) << " MB/s" << std::endl;
-    }
-    for(int block_size = 32; block_size <= 8192; block_size *= 2)
-    {
-        std::vector<char> buf(block_size, ' ');
-        int size = 0;
-        tmp.rewind();
-        boost::chrono::high_resolution_clock::time_point t1 = boost::chrono::high_resolution_clock::now();
-        while(size < data_size)
-        {
-            tmp.read(&buf[0], block_size);
-            size += block_size;
-        }
-        boost::chrono::high_resolution_clock::time_point t2 = boost::chrono::high_resolution_clock::now();
-        double tm = boost::chrono::duration_cast<boost::chrono::milliseconds>(t2 - t1).count() * 1e-3;
-        std::cout << "   read block size " << std::setw(8) << block_size << " " << std::fixed << std::setprecision(3)
-                  << (data_size / 1024.0 / 1024 / tm) << " MB/s" << std::endl;
-    }
-    tmp.close();
-    std::remove(file);
-}
-
-void test_perf(char *file)
-{
-    test_io<io_stdio>(file, "stdio");
-    test_io<io_fstream<std::fstream> >(file, "std::fstream");
-    test_io<io_fstream<nw::fstream> >(file, "nowide::fstream");
-}
-
 void make_empty_file(const char *filepath)
 {
     nw::ofstream f(filepath, std::ios_base::out | std::ios::trunc);
@@ -558,7 +431,7 @@ void test_fstream(const char *filename)
     TEST(nw::remove(filename) == 0);
 }
 
-int main(int argc, char **argv)
+int main(int, char **argv)
 {
     const std::string exampleFilename = std::string(argv[0]) + "-\xd7\xa9-\xd0\xbc-\xce\xbd.txt";
     try
@@ -585,11 +458,6 @@ int main(int argc, char **argv)
         test_flush<std::ifstream, std::ofstream>(exampleFilename.c_str());
         std::cout << "Flush - Test" << std::endl;
         test_flush<nw::ifstream, nw::ofstream>(exampleFilename.c_str());
-
-        if(argc == 2)
-        {
-            test_perf(argv[1]);
-        }
     } catch(std::exception const &e)
     {
         std::cerr << e.what() << std::endl;
